Move ClientIpUtils to web as it does not depend on Servlet API
